requirements.txt
ftfy==4.4.3
jupyter==1.0.0
Cython==0.29.14
matplotlib==3.1.1
networkx==2.4
nltk==3.4.5
numba==0.46.0
numpy==1.17.2
pandas==0.25.3
python-dotenv==0.10.3
sklearn
spacy==2.2.3
transformers==2.1
tensorflow==1.14
tensorflow_probability==0.7.0
torch==1.3.1
umap-learn==0.3.10
chinese-whispers==0.6.1
ax-platform==0.1.8
hdbscan==0.8.24
imbalanced-learn==0.6.1
pyyaml==5.3
seaborn==0.10.0
